first pass at ssh routing
