+++ go.mod
@@ -12,5 +12,6 @@ require (
 
 require (
 	golang.org/x/net v0.17.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
 )
